requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
flake8==3.4.1
idna==2.5
mccabe==0.6.1
nose==1.3.7
oauthlib==2.0.2
pep8==1.7.0
pycodestyle==2.3.1
pyflakes==1.5.0
python-dotenv==0.6.5
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
websocket-client==0.44.0
